Merging with mainline yt tip

--HG--
branch : yt-3.0

+++ yt/utilities/lib/CICDeposit.pyx
@@ -34,22 +34,22 @@ from fp_utils cimport imax, fmax, imin, fmin, iclip, fclip
 def CICDeposit_3(np.ndarray[np.float64_t, ndim=1] posx,
                  np.ndarray[np.float64_t, ndim=1] posy,
                  np.ndarray[np.float64_t, ndim=1] posz,
-                 np.ndarray[np.float32_t, ndim=1] mass,
+                 np.ndarray[np.float64_t, ndim=1] mass,
                  np.int64_t npositions,
-                 np.ndarray[np.float32_t, ndim=3] field,
+                 np.ndarray[np.float64_t, ndim=3] field,
                  np.ndarray[np.float64_t, ndim=1] leftEdge,
                  np.ndarray[np.int32_t, ndim=1] gridDimension,
                  np.float64_t cellSize):
 
     cdef int i1, j1, k1, n
-    cdef double xpos, ypos, zpos
-    cdef double fact, edge0, edge1, edge2
-    cdef double le0, le1, le2
-    cdef float dx, dy, dz, dx2, dy2, dz2
+    cdef np.float64_t xpos, ypos, zpos
+    cdef np.float64_t fact, edge0, edge1, edge2
+    cdef np.float64_t le0, le1, le2
+    cdef np.float64_t dx, dy, dz, dx2, dy2, dz2
 
-    edge0 = (<float> gridDimension[0]) - 0.5001
-    edge1 = (<float> gridDimension[1]) - 0.5001
-    edge2 = (<float> gridDimension[2]) - 0.5001
+    edge0 = (<np.float64_t> gridDimension[0]) - 0.5001
+    edge1 = (<np.float64_t> gridDimension[1]) - 0.5001
+    edge2 = (<np.float64_t> gridDimension[2]) - 0.5001
     fact = 1.0 / cellSize
 
     le0 = leftEdge[0]
@@ -68,9 +68,9 @@ def CICDeposit_3(np.ndarray[np.float64_t, ndim=1] posx,
         k1  = <int> (zpos + 0.5)
 
         # Compute the weights
-        dx = (<float> i1) + 0.5 - xpos
-        dy = (<float> j1) + 0.5 - ypos
-        dz = (<float> k1) + 0.5 - zpos
+        dx = (<np.float64_t> i1) + 0.5 - xpos
+        dy = (<np.float64_t> j1) + 0.5 - ypos
+        dz = (<np.float64_t> k1) + 0.5 - zpos
         dx2 =  1.0 - dx
         dy2 =  1.0 - dy
         dz2 =  1.0 - dz
